[mlx] fixes to improve mlx uses
