+++ Makefile
@@ -58,8 +58,10 @@ MANDATORY_FILES =	${CHECKS_PATH}/00_errors.c			\
 					${WIND}/key_handler.c				\
 					${WIND}/image.c						\
 					${WIND}/draw.c						\
+					${WIND}/temp.c						\
 					mainfer.c							\
 					exit.c								\
+					start_cub.c							\
 					# main.c							\
 
 #headers
