+++ sources/mlx_util/image.c
@@ -6,11 +6,19 @@
 /*   By: feralves <feralves@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/08 17:57:01 by feralves          #+#    #+#             */
-/*   Updated: 2023/08/09 20:45:08 by feralves         ###   ########.fr       */
+/*   Updated: 2023/08/12 11:03:23 by feralves         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
+
+int	render(t_vars *vars)
+{
+	draw_background(vars);
+	creating_img(vars);
+	mlx_put_image_to_window(vars->mlx, vars->win, vars->img.ptr, 0, 0);
+	return (0);
+}
 
 t_image	create_var_image(void *addr)
 {
